pyproject.toml
[project]
name = "analogue-weather-analysis"
version = "0.1.0"
description = "Analogue weather analysis following ClimaMeter/Ferranti methodology"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "xarray>=0.19.0",
    "netCDF4>=1.5.0",
    "dask>=2021.6.0",
    "pyyaml>=5.4.0",
]

[project.optional-dependencies]
viz = [
    "matplotlib>=3.4.0",
    "cartopy>=0.19.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["Python*"]
